Fix tests

+++ tests/test_main.py
@@ -1,11 +1,16 @@
-from tqdm import tqdm
+import os.path
 import shutil
-shutil.rmtree(".cache")
+
+from tqdm import tqdm
+
+if os.path.exists(".cache"):
+    shutil.rmtree(".cache")
+
 from ascl_net_scraper import __version__, scrape_details, scrape_index
 
 
 def test_main() -> None:
-    records = scrape_index(10)
+    records = scrape_index(20)
     for record in tqdm(records, total=len(records)):
         detailed_record = scrape_details(record.details_url)
         assert record.ascl_id == detailed_record.ascl_id
@@ -13,4 +18,4 @@ def test_main() -> None:
         assert record.credit == detailed_record.credit
         assert record.abstract == detailed_record.abstract
         assert record.details_url == detailed_record.url
-        detailed_record.github
+        assert isinstance(detailed_record.github, (str, type(None)))
